requirements/production.txt
gunicorn
Django==1.9.4
python-social-auth
requests
django-summernote
Pillow
celery[redis]
redis
honcho
psycopg2
dj-database-url
django-pipeline
djangorestframework
